requirements.txt
streamlit
google-generativeai
gTTS
python-dotenv
SpeechRecognition